backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
pymongo==4.6.1
python-decouple==3.8
razorpay==1.4.1
# sendgrid==6.10.0  # Removed - now using Gmail SMTP
reportlab==4.0.7
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
